Add "--compiler" command-line flag to specify the target compiler.
If no --compiler flag is given, cblrpm guesses the target compiler from the
$PATH environment variable, like before, but use e.g. --compiler=ghc-8.0.1
overrides that guess to provide deterministic output. This features is useful,
for example, when generating spec files targeting a compiler that the machine
running cblrpm does not have installed.

Note that Cabal is very liberal with regard to what a valid compiler ID is,
i.e. "--compiler=this-is-my-made-up-compiler" works just fine -- it's just not
going to produce meaningful spec files for all packages. The command line
parser will fail only if the given compiler id is syntactically incorrect, like
--compiler=-ghc-8.