If a library does not expose any modules, do not run haddock.
This comes up with the nats package.  This package has been merged into
base, but some things still require it via cabal.  If nats detects it is
building on a system where it is already a part of base, it will expose
no modules and attempting to run cabal-rpm will fail when it gets to
running haddock.

The fix is to make sure %ghc_lib_build_without_haddock is emitted in
this case.